+++ omicverse/bulk2single/__init__.py
@@ -4,4 +4,5 @@ Bulk2 single (A omic framework for bulk2single cell omic analysis)
 
 from ._bulk2single import Bulk2Single
 from ._single2spatial import Single2Spatial
+from ._bulktrajblend import BulkTrajBlend
 from ._utils import bulk2single_plot_cellprop,bulk2single_plot_correlation